% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_all_na.R
\name{alert_all_na}
\alias{alert_all_na}
\title{all_na_warning}
\usage{
alert_all_na(x, alert = base::warning)
}
\arguments{
\item{x}{argument}

\item{alert}{function such as \code{\link[base:worn]{base::worn()}} or \code{\link[base:message]{base::message()}} that gets
passed a character message.}
}
\description{
Simply produce an alert stating all values are missing and returns it's
argument invisibly. This does not actual test the argument, but is useful in
a test for missing values:
}
\details{
\preformatted{ if( all.na(x) ) return( all_na_warning(x) ) 
}
}
