% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.na.R
\name{all.na}
\alias{all.na}
\alias{all.na.default}
\alias{all.na.data.frame}
\alias{na.all}
\title{all.na, na.all}
\usage{
all.na(x)

\method{all}{na.default}(x)

\method{all}{na.data.frame}(x)

na.all(x)
}
\arguments{
\item{x}{object to test.}

\item{x}{any object}
}
\value{
logical scalar indicating if all values are missing.
}
\description{
Test if all values are missing
}
\details{
Implemented as:

For \strong{vectors}:\preformatted{ `. %>% is.na %>% all``
}

For \strong{data.frames}:\preformatted{ `. %>% sapply( all.na ) %>% all`` 
}

\code{na.all} is an alias of \code{all.na}. It is given as consistency with the other
function. \code{all.na} is preferred since the \code{na.*} functions are typically used
for vector imputaion.`
}
\examples{

  all.na( c( NA, NA, 1 ) )    # FALSE
  all.na( c( NA, NA, NA ) )   # TRUE
  
  df <- data.frame( char = rep(NA_character_, 3), nums=1:3)
  all.na(df)  # FALSE
  
  df <- data.frame( char = rep(NA_character_, 3), nums=rep(NA_real_,3))
  all.na(df)  # TRUE
  
}
\seealso{
\itemize{
\item \code{\link[=any.na]{any.na()}}
\item \code{\link[=which.na]{which.na()}}
}
}
