% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_na.R
\name{all_na}
\alias{all_na}
\alias{all_na.data.frame}
\alias{all_na.default}
\title{Determine if all values are mission}
\usage{
all_na(x)

\method{all_na}{default}(x)

\method{all_na}{data.frame}(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values in a data structure are missing
}
\details{
Implemented as: . %>% is.na %>% all
}
\examples{
all_na( c( NA, NA, 1 ) ) # FALSE
  all_na( c( NA, NA, NA ) )   # TRUE
}
\seealso{
\code{\link{all_na}} \cr
  \code{\link{which_na}}
}

