% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any.na.R
\name{any.na}
\alias{any.na}
\alias{any.na.default}
\alias{any.na.data.frame}
\alias{na.any}
\title{Determine if ANY value is NA}
\usage{
any.na(x)

\method{any.na}{default}(x)

\method{any.na}{data.frame}(x)

na.any(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values of an object is missing
}
\details{
For \strong{vectors}:\preformatted{     . %>% anyNA
```

For **tables**: 
}\preformatted{ . %>% is.na %>% any %>% any 
}\preformatted{`
Reports if any column has an `NA`,
}
}
\examples{
  any.na( 1:10 )           # FALSE
  any.na( c( 1, NA, 3 ) )  # TRUE
  
}
\seealso{
\itemize{
\item \code{\link[=base:anyNA]{base:anyNA()}}
\item \code{\link[=all.na]{all.na()}}
\item \code{\link[=which.na]{which.na()}}
}
}
