% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/any_na.R
\name{any_na}
\alias{any_na}
\alias{any_na.data.frame}
\alias{any_na.default}
\title{Determine if values are mission}
\usage{
any_na(x)

\method{any_na}{default}(x)

\method{any_na}{data.frame}(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values in a data structure are missing
}
\examples{
any_na( 1:10 )           # FALSE
  any_na( c( 1, NA, 3 ) )  # TRUE
}

