% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R, R/drop_na_rows.R
\name{drop_na_cols}
\alias{drop_na_cols}
\alias{drop_na_rows}
\title{Drop cols of a table whose values are all NA}
\usage{
drop_na_cols(data)

drop_na_rows(data)
}
\arguments{
\item{data}{data-like object

Remove all cols whose only values are NA.}

\item{data}{data-like object

Remove all rows whose only values are NA.}
}
\value{
An object of the same class as \code{data} with cols containing all 
  \code{NA} values removed

An object of the same class as \code{data} with rows containing all 
  \code{NA} values removed
}
\description{
Drop cols of a table whose values are all NA

Drop rows of a table whose values are all NA
}

