% Generated by roxygen2 (4.0.2): do not edit by hand
\name{howmany_na}
\alias{howmany_na}
\title{Counts how many values are NA}
\usage{
howmany_na(x)
}
\arguments{
\item{x}{object to count how many values are \code{NA}}
}
\description{
Returns the number of values that are NA
}
\examples{
x <- c( 1, NA, NA, 3:6 )
}
\keyword{manip}

