% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_zero}
\alias{impute_zero}
\alias{impute_zero_at}
\alias{impute_zero_all}
\alias{impute_zero_if}
\alias{impute_inf}
\alias{impute_inf_at}
\alias{impute_inf_all}
\alias{impute_inf_if}
\title{Table imputation methods}
\usage{
impute_zero(.tbl, ...)

impute_zero_at(.tbl, .vars, ...)

impute_zero_all(.tbl, ...)

impute_zero_if(.tbl, .na, .predicate, ...)

impute_inf(.tbl, ...)

impute_inf_at(.tbl, .vars, ...)

impute_inf_all(.tbl, ...)

impute_inf_if(.tbl, .na, .predicate, ...)
}
\arguments{
\item{.tbl}{table-like or list-like structure}

\item{...}{addition passed to the imputation method}

\item{.vars}{list of columns generated by vars(), or a character vector of
column names, or a numeric vector of column positions.}

\item{.predicate}{A predicate function to be applied to the columns or a
logical vector.}
}
\description{
Replace missing value methods with a variety of methods
}
\details{
These methods are modelled closely after \code{\link[dplyr:mutate]{dplyr::mutate()}} and related
functions
}
\examples{

  nacars \%>\% impute_zero() 
  nacars \%>\% impute_zero( mpg, cyl )
  nacars \%>\% impute_zero( "mpg", "cyl" )
  nacars \%>\% impute_zero( c("mpg","cyl") )
  nacars \%>\% impute_zero( 1:2 )
  

  nacars \%>\% impute_inf() 
  nacars \%>\% impute_inf( mpg, cyl )
  nacars \%>\% impute_inf( "mpg", "cyl" )
  nacars \%>\% impute_inf( c("mpg","cyl") )
  nacars \%>\% impute_inf( 1:2 )
  
}
