% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-recursive.R
\name{mk_recursive}
\alias{mk_recursive}
\title{mk_recursive}
\usage{
mk_recursive(fun)
}
\arguments{
\item{fun}{vector imputation method.}
}
\description{
Make table/recursive versions of vector replacement functions
}
\details{
Meta function that creates a table ready version of an vector replacement
method.
}
