% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-recursive.R
\name{na_constant}
\alias{na_constant}
\alias{na_zero}
\alias{na_inf}
\alias{na_neginf}
\title{Table imputation methods}
\usage{
na_constant(.tbl, ...)

na_zero(.data, ...)

na_inf(.tbl, ...)

na_neginf(.tbl, ...)
}
\arguments{
\item{.tbl}{table-like or list-like structure}

\item{...}{listname=expr}
}
\description{
Replace missing value methods
}
\details{
These methods perform missing value replacement on tables and lists.
}
