% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.all.R
\name{na.all}
\alias{na.all}
\alias{na.all.default}
\alias{na.all.data.frame}
\alias{all.na}
\title{na.all, all.na}
\usage{
na.all(x)

\method{na.all}{default}(x)

\method{na.all}{data.frame}(x)

all.na(x)
}
\arguments{
\item{x}{object to test.}

\item{x}{any object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Test if all values are missing
}
\details{
Implemented as:

For \strong{vectors}:\preformatted{ `. %>% is.na %>% all``
}

For \strong{data.frames}:\preformatted{ `. %>% sapply( na.all ) %>% any`` 
}

That is, reports if any columns are have all na's. This may change.

\code{all.na} is an alias of \code{na.all}.
}
\examples{
  na.all( c( NA, NA, 1 ) ) # FALSE
  na.all( c( NA, NA, NA ) )   # TRUE
  
  df <- data.frame( char = rep(na.character_, 3), nums=1:3)
  na.all(df)  # FALSE
  
  df <- data.frame( char = rep(na.character_, 3), nums=rep(na.real_,3))
  na.all(df)  # TRUE
  
}
\seealso{
\itemize{
\item \code{\link[=na.any]{na.any()}}
\item \code{\link[=na.which]{na.which()}}
}
}
