% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.any.R
\name{na.any}
\alias{na.any}
\alias{na.any.default}
\alias{na.any.data.frame}
\alias{any.na}
\title{Determine if ANY value is NA}
\usage{
na.any(x)

\method{na.any}{default}(x)

\method{na.any}{data.frame}(x)

any.na(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values of an object is missing
}
\details{
For \strong{vectors}:\preformatted{     . %>% anyNA
```

For **tables**: 
}\preformatted{ . %>% is.na %>% any %>% any 
}\preformatted{`
Reports if any column has an `NA`,
}
}
\examples{
  na.any( 1:10 )           # FALSE
  na.any( c( 1, NA, 3 ) )  # TRUE
  
}
\seealso{
\itemize{
\item \code{\link[=base:anyNA]{base:anyNA()}}
\item \code{\link[=na.all]{na.all()}}
\item \code{\link[=which.na]{which.na()}}
}
}
