% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.constant.R
\name{na.constant}
\alias{na.constant}
\title{na.constant}
\usage{
na.constant(x, .na)
}
\arguments{
\item{x}{vector; of values to have the NA replaced with \code{.na}}

\item{.na}{scalar .na to use as replacement.}
}
\value{
A vector with the type and length of \code{x} with all missing values replaces
by \code{.na}.
}
\description{
Replaces \code{NA}s by a constant
}
\details{
**This function is deprecated. Use \code{\link[=na.replace]{na.replace()}} instead.

Replaces \code{NA} values in the supplied vector with a default, \code{.na}.
}
\examples{
  na.constant( c(1,NA,2), -1 )

  na_constant <- mk_recursive(na.constant)
 
  mtcars \%>\%  na_constant( mpg=5, cyl=3 )
}
\seealso{
\code{\link[=na_replace]{na_replace()}}
\code{\link[=na_zero]{na_zero()}}  - a special case.
}
