% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.implicit.R
\name{na.implicit}
\alias{na.implicit}
\title{na.implicit}
\usage{
na.implicit(x, values)
}
\arguments{
\item{x}{object}

\item{values}{values to be recognized as NA}
}
\description{
Change certain values to NAs, ie make explicit \code{NAs} back to \code{NA}
}
\details{
\code{na.implicit} replaces \code{values} by \code{NA}. It is meant to be nearly inverse
operation to \code{na_explicit} (and \code{na_replace}). It can be used on both atomic
and recursive objects.

\code{na.implicit} is a S3 methods that can be used to defince additional
methods.
}
\examples{

 na.implicit( c(1,2,3,4), 3 )
 na.implicit( c("A", "NA", "B", "C") )
 na.implicit( c("A", "(Missing)", "B", "C") )
 
 df <- data.frame( char=c('A', 'NA', 'C', "(Missing)"), num=1:4 ) 
 na.implicit(df)
 
 
}
\seealso{
\itemize{
\item \code{na_explicit}
\item \code{na_replace}
}
}
