% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.inf.R
\name{na.inf}
\alias{na.inf}
\alias{na.neginf}
\title{Replace NAs with Inf}
\usage{
na.inf(x)

na.neginf(x)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by \code{Inf}
}
\details{
This is simply a wrapper aroung \code{\link{na.replace}} that
replaces \code{NA} values in the supplied vector with \code{Inf}
a default, \code{value}.
}
\examples{
na.inf( c( 1, 2, NA, 4) )
  na.neginf( c( 1, 2, NA, 4) )
}
\seealso{
\code{\link{na.replace}}
}

