% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.loess.R
\name{na.loess}
\alias{na.loess}
\title{na.loess}
\usage{
na.loess(object, ...)
}
\arguments{
\item{object}{Object to have \code{NA}'s replaced}

\item{...}{Additional arguments passed to loess

Replaces \code{NA} with a loess smoothed fit.  Cannot handle points at the
begginning or end of the series.}
}
\value{
Object of the same type with NAs replaced by a loess smoothed line
}
\description{
Replace \code{NA} with loess model
}
\note{
This does not follow function naming conventions in order to more closely
resembling existing \code{na.*} functions
}
\examples{
  
  x <- 1:101 
  x[ seq(20,80,5)] <- NA
  na.loess(x)

}
