% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.max.R
\name{na.max}
\alias{na.max}
\title{na.max, na.min}
\usage{
na.max(x, ...)
}
\arguments{
\item{x}{vector to have mising values replaced.}

\item{...}{does nothing

\code{na.max} and \code{na.min} replace missing values (\code{NA}) with the maximum or
minimum of non-missing values \code{x}. Internally \code{\link[base:max]{base::max()}} and \code{\link[base:min]{base::min()}}
with \code{na.rm=TRUE}.

\code{...} passes additional arguments to \code{\link[base:max]{base::max()}} and \code{\link[base:min]{base::min()}} which
do nothing.}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the maximum value
of \code{x}; missing values are omitted.
}
\description{
Replace \code{NA} values with the max/min in a safe, consitent way.
}
\seealso{
\itemize{
\item \code{\link[=na.replace]{na.replace()}} - used internally by these functions
\item \code{\link[base:max]{base::max()}} and \code{\link[base:min]{base::min()}}
\item \code{\link[=na.inf]{na.inf()}} / \link{na.neginf}
}
}
