% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.mean.R
\name{na.mean}
\alias{na.mean}
\title{na.mean}
\usage{
na.mean(x, ...)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.}

\item{...}{additional arguments passed to \code{\link[=base:mean]{base:mean()}} such as

\code{na.mean} replaces \code{NA} values with the mean of \code{x}. Internally,
\code{mean(x, na.rm=TRUE, ... )} is used. If mean cannot be calculated (e.g. x
isn't numeric) then \code{x} is returned with a warning.}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the \code{mean(x)}
}
\description{
Replace \code{NA} values with the mean
}
\seealso{
\code{\link[=na.replace]{na.replace()}}
\code{\link[=na.median]{na.median()}}`
}
