% Generated by roxygen2 (4.0.1): do not edit by hand
\name{na.min}
\alias{na.min}
\title{na.min}
\usage{
na.min(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.


Replaces \code{NA} values with the min of the series. Will return NA if
all the values are \code{NA}}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the min
}
\description{
Replace \code{NA} values with the min
}
\seealso{
\code{\link{na.replace}}, \code{\link{na.max}}
}

