% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.most_freq.R
\name{na.most_freq}
\alias{na.most_freq}
\title{na.most_freq}
\usage{
na.most_freq(x, .na, ...)
}
\arguments{
\item{...}{arguments passed to functions}
}
\description{
Replace missing value with the most frequently-occuring non-missing value
}
\details{
In the event of ties, the values listed first will be used.
}
\examples{

  c(1,1,NA,4) \%>\% na.most_freq()
  c(1,1,4,4,NA) \%>\% na.most_freq() 
  
}
