% Generated by roxygen2 (4.0.1): do not edit by hand
\name{na.nocb}
\alias{na.nocb}
\title{na.nocb (_N_ext _O_bservation _C_arried _B_ack )}
\usage{
na.nocb(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{arguments passed to \code{na.locf}

\code{na.nocb} uses \code{zoo::}\code{na.locf}.}
}
\value{
a object with NA's replaced by the next occuring value in the sequence
}
\description{
na.nocb (_N_ext _O_bservation _C_arried _B_ack )
}
\seealso{
\code{\link[zoo]{na.locf}}
}

