% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.predict.R
\name{na.predict}
\alias{na.predict}
\title{na.predict}
\usage{
na.predict(x, object, data = x)
}
\arguments{
\item{x}{data}

\item{object}{object with predict method}

\item{data}{data object}
}
\description{
replace \code{NA} values by predictions of a model
}

