% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.rm.R
\name{na.rm}
\alias{na.rm}
\alias{na_rm}
\title{na.rm}
\usage{
na.rm(object)

na.rm(object)

na_rm(object)
}
\arguments{
\item{object}{to remove \code{NA}s from}
}
\value{
An object of the same class with all \code{NA} values removed. For
data.frame and data.table objects entire columns are removed if they
contain solely \code{NA} values.
}
\description{
Removes \code{NA} values from objects
}
\details{
For \strong{vectors} this is the same as \code{\link[stats:na.omit]{stats::na.omit()}} or
\code{\link[stats:na.exclude]{stats::na.exclude()}}.

For removing columns or rows that are have only missing values use
\code{\link[=na.drop_col]{na.drop_col()}} or \code{\link[=na.drop_row]{na.drop_row()}}
}
\seealso{
\itemize{
\item \code{\link[stats:na.omit]{stats::na.omit()}}, \code{\link[stats:na.exclude]{stats::na.exclude()}}
\item \code{\link[=drop_cols_all_na]{drop_cols_all_na()}}
\item \code{\link[=all.na]{all.na()}}
}
}
