% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.roughfix.POSIXct.R
\name{na.roughfix.POSIXct}
\alias{na.roughfix.POSIXct}
\title{na.roughfix.POXIXct - replace a date/time field}
\usage{
na.roughfix.POSIXct(object)
}
\arguments{
\item{object}{to replace \code{NA}s with}
}
\description{
na.roughfix.POXIXct - replace a date/time field
}

