% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.roughfix.POSIXct.R
\name{na.roughfix.POSIXct}
\alias{na.roughfix.POSIXct}
\title{na.roughfix.POXIXct}
\usage{
na.roughfix.POSIXct(object)
}
\arguments{
\item{object}{to replace missing values with.}
}
\description{
Replace a date/time field
}
\details{
\code{na.roughfix.POSIXct} implements a POSIXct method for
\code{na.roughfix}from the \emph{randomForest} package. This violates the
}
