% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.sample.R
\name{na.sample}
\alias{na.sample}
\alias{na.random}
\title{na.sample}
\usage{
na.sample(x, population = na.omit(x), replace = TRUE, prob = NULL)

na.random(x, ...)
}
\arguments{
\item{x}{vector with}

\item{population}{ither a vector of one or more elements from which to choose, or a
positive integer. See \code{\link[base:sample]{base::sample()}}. (Default: \code{na.omit(x)})}

\item{replace}{logical indicationg that sampling should be performed
with replacement? (Default: `TRUE``)}

\item{prob}{vector of probability weights for obtaining the elements of
the vector being sampled. (Default: \code{NULL} equal weights)}
}
\description{
Replace missing values with value randomly drawn from x
}
\details{
\code{na.random} replaces missing values by sampling the non-missing values
\strong{with replacement}. This relies on \code{\link[base:sample]{base::sample()}}.

The default is to replace bv sampling a population defined by the
non-missing values of \code{.x} \strong{with replacement}

\code{na.random} is an alias for \code{na.sample}vlu
`
}
\note{
\code{na.sample} is \strong{non-deterministic}. Use
\code{\link[base:set.seed]{base::set.seed()}} to make it deterministic
}
