% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.which.R
\name{na.which}
\alias{na.which}
\alias{na.which.default}
\alias{na.which.data.frame}
\alias{which.na}
\title{na.which - Identify missing values}
\usage{
na.which(x)

\method{na.which}{default}(x)

\method{na.which}{data.frame}(x)

which.na(x)
}
\arguments{
\item{x}{vector to search for missing values}
}
\value{
\code{integer} of indexes of \code{x} that are \code{NA}.  The
\code{names} of the returned vector are set to the names of \code{x}.
that are \code{NA}.
}
\description{
Return the indexes of those values that are missing
}
\details{
Implemented as \code{which( is.na(x) )}
}
\examples{
  x <- c( 1, NA, NA, 3:6 )
  na.which(x)
  
}
