% Generated by roxygen2 (4.0.1): do not edit by hand
\name{na.zero}
\alias{na.zero}
\title{na.zero}
\usage{
na.zero(x)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}
}
\value{
\code{x} with \code{NA}s replaced by 0.
}
\description{
Replaces \code{NA}s by 0.
}
\details{
Replaces \code{NA} values in the supplied vector with
a default, \code{value}.
}

