% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_any.R
\name{na_any}
\alias{any_na}
\alias{na_any}
\alias{na_any.data.frame}
\alias{na_any.default}
\title{Determine if ANY value is NA}
\usage{
na_any(x)

\method{na_any}{default}(x)

\method{na_any}{data.frame}(x)

any_na(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values of an object is missing
}
\details{
Implemented as: . %>% anyNA

For data.tables: . %>% is.na %>% any %>% any 

Reports if any column has an \code{NA},
}
\examples{
  na_any( 1:10 )           # FALSE
  na_any( c( 1, NA, 3 ) )  # TRUE
  
}
\seealso{
\code{\link[base]{anyNA}} \cr
  \code{\link{na_all}} \cr
  \code{which_na} \cr
}

