% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_constant.R
\name{na_constant}
\alias{na_constant}
\title{na_constant}
\usage{
na_constant(x, value)
}
\arguments{
\item{x}{vector; of values to have the NA replaced with \code{value}}

\item{value;}{single value to replace \code{NA}}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by a constant
}
\details{
Replaces \code{NA} values in the supplied vector with a default, 
\code{value}.
}
\examples{
  na_constant( c(1,NA,2), -1 )

}
\seealso{
\code{\link{na_replace}}  \cr
  \code{\link{na_zero}} - a special case \cr
}
