% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_drop_cols.R
\name{na_drop_cols}
\alias{na_drop_cols}
\alias{drop_na_cols}
\title{Drop cols of a table whose values are all NA}
\usage{
na_drop_cols(data)

drop_na_cols(data)
}
\arguments{
\item{data}{data-like object

Remove all cols whose only values are NA.}
}
\value{
An object of the same class as \code{data} with cols containing all 
  \code{NA} values removed
}
\description{
Drop cols of a table whose values are all NA
}
