% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_explicit_.R
\docType{data}
\name{NA_explicit_}
\alias{NA_explicit_}
\title{NA_explicit_}
\format{An object of class \code{NULL} of length 0.}
\usage{
NA_explicit_
}
\description{
Constant for categorical/character/factor variables
}
\details{
\code{NA_explicit_} is an active binding to \code{getOptions('NA_explicit_})\code{and allows for a consistent, explicit}NA` values.
}
\seealso{
\itemize{
\item \code{\link[=na_explicit]{na_explicit()}}
\item \code{\link[=na_replace]{na_replace()}}
}
}
\keyword{datasets}
