% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_fill.R
\name{na_fill}
\alias{na_fill}
\alias{na_fill.default}
\alias{na_fill.ts}
\alias{na_fill.zoo}
\title{Fill NA or specified positions.}
\usage{
na_fill(object, fill, ...)
}
\arguments{
\item{object}{an object.}

\item{fill}{a three component list or a vector that is coerced to a list.
Shorter objects are recycled.  The three components represent the fill value
to the left of the data, within the interior of the data and to the right of
the data, respectively.  The value of any component may be the keyword
\code{"extend"} to indicate repetition of the leftmost or rightmost non-NA
value or linear interpolation in the interior.  \code{NULL} means that items
are dropped rather than filled.}

\item{ix}{logical. Should be the same length as the number of time points.
Indicates which time points not to fill. This defaults to the non-NA values.}

\item{\dots}{further arguments passed to methods.}
}
\description{
Generic function for filling \code{NA} values or specified positions.
}
\details{
Fill \code{NA} or indicated values.
}
\examples{


z <- zoo(c(NA, 2, NA, 1, 4, 5, 2, NA))
na_fill(z, "extend")
na_fill(z, c("extend", NA))
na_fill(z, -(1:3))
na_fill(z, list(NA, NULL, NA))


}
\seealso{
\code{\link{na_approx}}
}
\keyword{ts}

