% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_implicit.R
\name{na_implicit}
\alias{na_implicit}
\title{na_implicit}
\usage{
na_implicit(x, values)
}
\arguments{
\item{x}{object}

\item{values}{values to be recognized as NA}
}
\description{
Change values (explicit `NAs`) back to `NA`
}
\details{
`na_implicit` replaces `values` by `NA`. It is meant to be nearly inverse 
 operation to `na_explicit` (and `na_replace`). It can be used on both atomic 
 and recursive objects.  
 
 `na_implicit` is a S3 methods that can be used to defince additional 
 methods.
}
\examples{

 na_implicit( c(1,2,3,4), 3 )
 na_implicit( c("A", "NA", "B", "C") )
 na_implicit( c("A", "(Missing)", "B", "C") )
 
 df <- data.frame( char=c('A', 'NA', 'C', "(Missing)"), num=1:4 ) 
 na_implicit(df)
 
 
}
\seealso{
- `na_explicit`
 - `na_replace`
}
