% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na_inf.R
\name{na_inf}
\alias{na_inf}
\alias{na_neginf}
\title{Replace NAs with Inf}
\usage{
na_inf(x)

na_neginf(x)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by \code{Inf}
}
\details{
This is simply a wrapper aroung \code{\link{na_replace}} that
replaces \code{NA} values in the supplied vector with \code{Inf}
a default, \code{value}.
}
\examples{
na_inf( c( 1, 2, NA, 4) )
  na_neginf( c( 1, 2, NA, 4) )
}
\seealso{
\code{\link{na_replace}}
}

