% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_max.R
\name{na_max}
\alias{na_max}
\title{na_max}
\usage{
na_max(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.


Replaces \code{NA} values with the max of the series. Will return NA if 
all the values are \code{NA}}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the max
}
\description{
Replace \code{NA} values with the max
}
\seealso{
\code{\link{na_replace}}, \code{\link{na_min}}
}

