% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_median.R
\name{na_median}
\alias{na_median}
\title{na_median}
\usage{
na_median(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the 
  median
}
\description{
Replace \code{NA} values with the median
}
\seealso{
\code{\link{na_replace}}, \code{\link{na_mean}}
}
