% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_min.R
\name{na_min}
\alias{na_min}
\title{na_min}
\usage{
na_min(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.


Replaces \code{NA} values with the min of the series. Will return NA if 
all the values are \code{NA}}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the min
}
\description{
Replace \code{NA} values with the min
}
\seealso{
\code{\link{na_replace}}, \code{\link{na_max}}
}

