% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_n.R
\name{na_n}
\alias{na_n}
\alias{na_howmany}
\title{Counts how many values are NA}
\usage{
na_n(x)

na_howmany(x)
}
\arguments{
\item{x}{object to count how many values are \code{NA}}
}
\description{
Returns the number of values that are NA
}
\examples{
  x <- c( 1, NA, NA, 3:6 )

}
