% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace.R
\name{na_replace}
\alias{na_replace}
\alias{na.replace}
\title{Replace NA valies}
\usage{
na_replace(x, value)

na.replace(x, value = 0)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}

\item{value}{to replace \code{NA}s}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by \code{value}
}
\details{
\code{na_replace} is an alias of: 

     ifelse( is.na(x), value, x )

It replaces \code{NA} values in the input vector with 
a default, \code{value}. \code{value} can either be a length-1 scalar or a 
vector of \code{length(x)}.  Recycling is not allowed.  If \code{value} is 
a scalar, all NA values are replaced by \code{value}.  If a vector, only 
\code{NA} of \code{x} are replaced by the corresponding items of \code{value}.

It is mainly provided as a cleaner, pipe-able replacement for ifelse.
}
\examples{

na_replace( c(1,NA,3,NA), 2 )    # 1 2 3 2   
na_replace( c(1,NA,3,NA), 1:4 )  # 1 2 3 4

}
\seealso{
\code{\link[base]{ifelse}}
}
