% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm.R
\name{na_rm}
\alias{na_rm}
\title{na_rm}
\usage{
na_rm(object)

na_rm(object)
}
\arguments{
\item{object}{to remove \code{NA}s from}
}
\value{
An object of the same class with all \code{NA} values removed. For
data.frame and data.table objects entire columns are removed if they
contain solely \code{NA} values.
}
\description{
Removes \code{NA} values from objects
}
\details{
For \strong{vectors} this is the same as \code{\link[stats:na_omit]{stats::na_omit()}} and
\code{\link[stats:na_exclude]{stats::na_exclude()}}.

For removing columns or rows that are have only missing values use
\code{\link[=na_drop_col]{na_drop_col()}} or \code{\link[=na_drop_row]{na_drop_row()}}
}
\seealso{
\code{\link[stats:na_omit]{stats::na_omit()}}
\code{\link[stats:na_exclude]{stats::na_exclude()}} \cr
\code{\link[=na_all]{na_all()}}
}
