% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_roughfix.POSIXct.R
\name{na_roughfix.POSIXct}
\alias{na_roughfix.POSIXct}
\title{na_roughfix.POXIXct - replace a date/time field}
\usage{
na_roughfix.POSIXct(object)
}
\arguments{
\item{object}{to replace \code{NA}s with}
}
\description{
na_roughfix.POXIXct - replace a date/time field
}
