% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_which.R
\name{na_which}
\alias{na_which}
\alias{na_which_default}
\alias{na_which_data.frame}
\alias{which_na}
\title{na_which: Identify missing values}
\usage{
na_which(x)

na_which_default(x)

na_which_data.frame(x)

which_na(x)
}
\arguments{
\item{x}{Vector to search for missing values}
}
\value{
\code{integer} of indexes of \code{x} that are \code{NA}.  The
\code{names} of the returned vector are set to the names of \code{x}.
that are \code{NA}.
}
\description{
Return the indexes of those values that are missing
}
\details{
Implemented as . %>% is.na %>% which 

For data.frames:
}
\examples{
  x <- c( 1, NA, NA, 3:6 )
  na_which(x)
}
