% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na_zero.R
\name{na_zero}
\alias{na_zero}
\title{na_zero}
\usage{
na_zero(x)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}
}
\value{
\code{x} with \code{NA}s replaced by 0.
}
\description{
Replaces \code{NA}s by 0.
}
\details{
Replaces \code{NA} values in the supplied vector with
a default, \code{0}.
}
\seealso{
\code{\link{na_replace}} \cr
  \code{\link{na_constant}}
}

