% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.constant.R
\name{na.constant}
\alias{na.constant}
\title{na.constant}
\usage{
na.constant(x, value)
}
\arguments{
\item{x}{vector; of values to have the NA replaced with \code{value}}

\item{value;}{single value to replace \code{NA}}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by a constant
}
\details{
Replaces \code{NA} values in the supplied vector with a default,
\code{value}.
}
\examples{
na.constant( c(1,NA,2), -1 )
}
\seealso{
\code{\link{na.replace}}  \cr
  \code{\link{na.zero}} - a special case \cr
}

