% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.max.R
\name{na.max}
\alias{na.max}
\title{na.max}
\usage{
na.max(x, ...)
}
\arguments{
\item{x}{vector to have mising values of \code{NA} values replaced.}

\item{...}{additional arguments passed to other functions

\code{na.max} replaces missing values (\code{NA}) with the maximum of non-missing
values \code{x} using \code{base::max(na.rm==TRUE)}. \code{x} has to be numeric. ...`
passes additional arguments to \code{\link[base:max]{base::max()}}.}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the maximum value
of \code{x}; missing values are omitted.
}
\description{
Replace \code{NA} values with the max
}
\seealso{
\code{\link[=na.replace]{na.replace()}} - used internally by these function
\code{\link[=na.min]{na.min()}} -
\code{\link[=na.inf]{na.inf()}} / \link{na.neginf}
}
