% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{na.max}
\alias{na.max}
\title{na.max}
\usage{
na.max(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.


Replaces \code{NA} values with the max of the series. Will return NA if
all the values are \code{NA}}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the max
}
\description{
Replace \code{NA} values with the max
}
\seealso{
\code{\link{na.replace}}, \code{\link{na.min}}
}

