% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.mean.R
\name{na.mean}
\alias{na.mean}
\title{na.mean}
\usage{
na.mean(x, ...)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.}

\item{...}{additional arguments passed to \code{\link[base]{mean}} such as


Replaces \code{NA} values with the mean of the series. Will return NA if
all the values are \code{NA}}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the mean
}
\description{
Replace \code{NA} values with the mean
}
\seealso{
\code{\link{na.replace}}, \code{na.median}
}

