% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.median.R
\name{na.median}
\alias{na.median}
\title{na.median}
\usage{
na.median(x)
}
\arguments{
\item{x}{numeric; vector to have \code{NA} values replaced.}
}
\value{
A vector of the same length as \code{x} with \code{NA} replaced by the
  median
}
\description{
Replace \code{NA} values with the median
}
\seealso{
\code{\link{na.replace}}, \code{\link{na.mean}}
}

