% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.n.R
\name{na.n}
\alias{na.n}
\alias{na.howmany}
\alias{na.pct}
\title{Counts how many values are NA}
\usage{
na.n(x)

na.howmany(x)

na.pct(x)
}
\arguments{
\item{x}{object to count how many values are \code{NA}}
}
\description{
Returns the number of values that are NA
}
\examples{
  x <- c( 1, NA, NA, 4:5 )
  na.n(x)
  na.pct(x)

}
