% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\alias{na_replace}
\title{Replace NA valies}
\usage{
na.replace(x, value = 0)

na_replace(x, value = 0)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}

\item{value}{to replace \code{NA}}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by \code{value}
}
\details{
Replaces \code{NA} values in the supplied vector with
a default, \code{value}.
}

