% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.rm.R
\name{na.rm}
\alias{na.rm}
\title{na.rm}
\usage{
na.rm(object)
}
\arguments{
\item{object}{to remove \code{NA}s from}
}
\value{
An object of the same class with all \code{NA} values removed. For
  data.frame and data.table objects entire columns are removed if they
  contain solely \code{NA} values.
}
\description{
Removes \code{NA} from objects
}
\details{
For \strong{vectors} this is the same as \code{\link[stats]{na.omit}} and
\code{\link[stats]{na.exclude}}.

For \strong{data.frames} and \strong{data.tables} columns that are entirely
\code{NA} are removed.
}
\seealso{
\code{\link[stats]{na.omit}} \cr
  \code{\link[stats]{na.exclude}} \cr
  \code{\link{all_na}}
}

