% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_all.R
\name{na_all}
\alias{na_all}
\alias{na_all.data.frame}
\alias{na_all.default}
\title{Determine if all values are mission}
\usage{
na_all(x)

\method{na_all}{default}(x)

\method{na_all}{data.frame}(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{logical}

\code{logical}
}
\description{
Determines if any of the values in a data structure are missing
}
\details{
Implemented as: . %>% is.na %>% all
}
\examples{
  na_all( c( NA, NA, 1 ) ) # FALSE
  na_all( c( NA, NA, NA ) )   # TRUE
  
}
\seealso{
\code{\link{na_all}} \cr
  \code{\link{which_na}}
}

