% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_explicit.R
\name{na_explicit}
\alias{na_explicit}
\title{na_explicit}
\usage{
na_explicit(x, values)
}
\arguments{
\item{x}{object; either atomic or recursive.}

\item{values}{either a single element list or named vectors; See #Details.}
}
\description{
Make NA explicit values
}
\details{
`na_explicit` transforms `NA` to values. It is like 
`na_replace` but provides an S3 interface that can provide specific 
 actions for various types/classes. It also works on recursive objects (i.e. 
 lists, data.frames, etc) by applying na_explicit to each element.
 
 It is also intended to be single argument function.
}
\examples{

   na_explicit( c(1, NA, 3, 4), 0)
   na_explicit( c("A",NA,"c","D" ) ) 

}
\seealso{
- na.actions::na_replace 
 - na.actions::na_implicit
 - na.actions::na_replace
 - forcats::fct_explicit_na
}
