% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace.R
\name{na_replace}
\alias{na_replace}
\alias{na_replace_at}
\title{Replace missing values in tbales and lists}
\usage{
`na_replace` can be used for three replacement operatations: 
 
 1. `na_replace( .data, .na )` : ( missing `...` ) Replace missing values 
     in **ALL COLS** by `.na`
 2. `na_replace( .data, .na, ... )` : ( `...` is an unnamed list) Replace 
     column(s) specified in `...` by `.na`.  Columns are specified as an 
     unnamed list of quoted or unquoted column names.
 3. `na_replace( .data. col1=na.*, col2=na.* )` : ( missing `.na` ) : 
    Replace by column-specific `.na` 
 
Use Additional arguments are to `.na` are not used; Use na_replace_at for 
or create your own lambda functions.

na_replace_at(.tbl, .na, .vars, ...)
}
\arguments{
\item{.data}{list-like or table-like structure.}

\item{.na}{scalar, vector or function as described in \code{\link[=na.replace]{na.replace()}}}

\item{...}{additional args; either a unnamed list of columns (quoted or not)
or name=function pairs.  See Details.}
}
\value{
Returns a object as the same type as \code{.data}. Columns are mutated to replace
missing values (\code{NA}) with value specied by \code{.na} and \code{...}
}
\description{
Replace missing values (\code{NA}) in a table and lists
}
\details{
\code{na_replace} is similar to other \emph{dplyr} verbs especially \code{\link[dplyr:mutate]{dplyr::mutate()}}
except it always operates on a subset of vector elements, i.e. those with \code{NA}
values.

The behavior depends on the values of \code{.na} and \code{...}. (See Usage.)
}
\note{
\code{...} is used to specify columns in \code{na_replace} but is used as additional
arguments to \code{.na} in the other \code{na_replace_*} functions.
}
\examples{

data(nacars)
nacars \%>\% na_replace(0, mpg, cyl)
nacars \%>\% na_replace(1:6, mpg, cyl)

nacars \%>\% na_replace( na.mean )
nacars \%>\% na_replace( mean )       # unsafe
nacars \%>\% na_replace( length, mpg, disp )
nacars \%>\% na_replace( mean, mpg, disp )
nacars \%>\% na_replace( mpg=na.mean, cyl=na.max )
nacars \%>\% na_replace( na.mean, c('mpg','disp') )
  

 nacars \%>\% na_replace_at( -99, .vars=1:3 )
 nacars \%>\% na_replace_at( .na=na.mean, .vars=1:6, trim=.25 )  # Uses `...` for additional args
 
}
