% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace.R
\name{na_replace}
\alias{na_replace}
\alias{na.replace}
\title{Replace NA valies}
\usage{
na_replace(x, value)

na.replace(x, value = 0)
}
\arguments{
\item{x}{vector of values to have the NA replaced with \code{value}}

\item{value}{to replace \code{NA}s}
}
\value{
\code{x} with \code{NA}s replaced by \code{value}.
}
\description{
Replaces \code{NA}s by \code{value}
}
\details{
\code{na_replace} replaces \code{NA} values in the input vector with 
a default, \code{value}. \code{value} can either be a length-1 scalar or a 
vector of \code{length(x)}.  If a scalar, all NA values are replaced by the 
scalar value.  If a vector, only \code{NA} values in \code{x} are replaced.
}
\examples{

na_replace( c(1,NA,3), 2 )    # 1 2 3   
na_replace( c(1,NA,3), 1:3 )  # 1 2 3



}
