% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace.R
\name{na_replace}
\alias{na_replace}
\title{Replace missing values in tbales and lists}
\usage{
`na_replace` can be used for three replacement operatations: 
 
 1. `na_replace( .data, .na )` : ( missing `...` ) Replace missing values 
     in **ALL COLS** by `.na`
 2. `na_replace( .data, .na, ... )` : ( `...` is an unnamed list) Replace 
     column(s) specified in `...` by `.na`.  Columns are specified as an 
     unnamed list of quoted or unquoted column names.
 3. `na_replace( .data. col1=na.*, col2=na.* )` : ( missing `.na` ) : 
    Replace by column-specific `.na` 
 
Use Additional arguments are to `.na` are not used; Use na_replace_at for 
or create your own lambda functions.
}
\arguments{
\item{.data}{list-like or table-like structure.}

\item{.na}{scalar, vector or function as described in \code{\link[=na.replace]{na.replace()}}}

\item{...}{additional args; either a unnamed list of columns (quoted or not)
or name=function pairs.  See Details.}
}
\value{
Returns a object as the same type as \code{.data}. Columns are mutated to replace
missing values (\code{NA}) with value specied by \code{.na} and \code{...}
}
\description{
Replace missing values (\code{NA}) in a table and lists
}
\details{
\code{na_replace} is similar to other \emph{dplyr} verbs especially \code{\link[dplyr:mutate]{dplyr::mutate()}}
except it always operates on a subset of vector elements, i.e. those with \code{NA}
values.

The behavior depends on the values of \code{.na} and \code{...}. (See Usage.)
}
\examples{

mtcars \%>\% na_replace(0, mpg, cyl)
mtcars \%>\% na_replace(1:6, mpg, cyl)

mtcars \%>\% na_replace( na.mean )
mtcars \%>\% na_replace( mean )       # unsafe
mtcars \%>\% na_replace( length, mpg, disp )
mtcars \%>\% na_replace( mean, mpg, disp )
mtcars \%>\% na_replace( mpg=na.mean, cyl=na.max )

}
