% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.na.R
\name{which.na}
\alias{which.na}
\alias{which.na.default}
\alias{which.na.data.frame}
\alias{na.which}
\title{which.na - Identify missing values}
\usage{
which.na(x)

\method{which.na}{default}(x)

\method{which.na}{data.frame}(x)

na.which(x)
}
\arguments{
\item{x}{vector to search for missing values}
}
\value{
\code{integer} of indexes of \code{x} that are \code{NA}.  The
\code{names} of the returned vector are set to the names of \code{x}.
}
\description{
Return the indexes of those values that are missing
}
\details{
\code{which.na} is implemented as \code{which( is.na(x) )} and
}
\examples{

  x <- c( 1, NA, NA, 4:6 )
  which.na(x)
  
  names(x) <- letters[1:6]
  which.na(x)
  
}
\seealso{
\itemize{
\item \code{\link[base:is.na]{base::is.na()}} - for the variant returning logical
}
}
