% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/which_na.R
\name{which_na}
\alias{which_na}
\alias{which_na.data.frame}
\alias{which_na.default}
\title{which_na: Identify missing values}
\usage{
which_na(x)

\method{which_na}{default}(x)

\method{which_na}{data.frame}(x)
}
\arguments{
\item{x}{Vector to search for missing values}
}
\value{
\code{integer} of indexes of \code{x} that are \code{NA}.  The
\code{names} of the returned vector are set to the names of \code{x}.
that are \code{NA}.
}
\description{
Return the indexes of those values that are missing
}
\examples{
x <- c( 1, NA, NA, 3:6 )
  which_na(x)
}

